% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5_snowtrends_map_hourly_functions.R
\name{appendvar}
\alias{appendvar}
\title{Append variable from one data frame to another}
\usage{
appendvar(var1, var2, var3)
}
\arguments{
\item{var1}{Vector of values to match}

\item{var2}{Vector to match against}

\item{var3}{Vector of values to return}
}
\value{
Vector of matched values
}
\description{
Utility function for matching and appending variables between data frames.
This function matches values in var1 to var2 and returns corresponding
values from var3.
}
\examples{
# Match site names to get coordinates
sites <- c("Site1", "Site2", "Site3")
site_names <- c("Site1", "Site2", "Site3", "Site4")
lng <- c(-120, -121, -122, -123)

longitude <- appendvar(sites, site_names, lng)
}

